requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-decouple==3.8
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
httpx>=0.25.2,<0.26.0
redis==5.0.1
ollama==0.1.7
python-dotenv==1.0.0
prometheus-client==0.19.0